+++ lib/jlc-parts-engine/jlc-parts-cache.ts
@@ -1,0 +1,26 @@
+export const cache = new Map<string, any>()
+
+export const getJlcPartsCached = async (name: any, params: any) => {
+  const paramString = new URLSearchParams({
+    ...params,
+    json: "true",
+  }).toString()
+
+  if (cache.has(paramString)) {
+    return cache.get(paramString)
+  }
+
+  const response = await fetch(
+    `https://jlcsearch.tscircuit.com/${name}/list?${paramString}`,
+  )
+  const responseJson = await response.json()
+  cache.set(paramString, responseJson)
+  return responseJson
+}
+
+export const withBasicPartPreference = (parts: any[] | undefined) => {
+  if (!parts) return []
+  return [...parts].sort(
+    (a, b) => Number(b.is_basic ?? false) - Number(a.is_basic ?? false),
+  )
+}
