+++ lib/jlc-parts-engine/types.ts
@@ -1,0 +1,11 @@
+export type PlatformFetch = typeof fetch
+
+export type FetchPartCircuitJsonParams = {
+  supplierPartNumber?: string
+  manufacturerPartNumber?: string
+  platformFetch?: PlatformFetch
+}
+
+export type CreateJlcPartsEngineOptions = {
+  platformFetch?: PlatformFetch
+}
